\name{config.summary}
\alias{config.summary}
\title{Print summary of configuration}
\usage{
config.summary(config.R, indent = "  ")
}
\arguments{
  \item{config.R}{path to configuration file}

  \item{indent}{level for indenting output before required
  input file}
}
\description{
This function outputs the summary of the required files in
the configuration file.
}

