\name{load.config}
\alias{load.config}
\title{Load the configuration file}
\usage{
  load.config(config.R)
}
\arguments{
  \item{config.R}{path to configuration file}
}
\description{
  This function loads the specified configuration file into
  the global namespace. [TODO: specify needed config.R
  format in seperate help file.]
}
\examples{
load.config('/usr/local/ohi/src/toolbox/scenarios/global_2012_nature/conf/config.R')
}
\keyword{app}

