\name{ohi.extras}
\alias{apply.bycol}
\alias{apply.byrow}
\alias{ends.with}
\alias{mangle}
\alias{md.table}
\alias{ohi.extras}
\alias{ohi.find.file}
\alias{ohi.goal.to.old.goal.component}
\alias{ohi.model.keys}
\alias{ohi.old.goal.component}
\alias{ohi.old.goal.component.to.goal}
\alias{ohi.options}
\alias{ohi.write.results}
\alias{pkey}
\alias{scenario.to.beta}
\alias{scenarios.cases}
\alias{scenarios.compare}
\alias{starts.with}
\alias{vcat}
\alias{zstats.load}
\title{Ocean Health Index: TBD}
\source{
tbd
}
\description{
TBD
}
\examples{
Sys.info()
}
\references{
tbd
}

