\docType{package}
\name{ohigui-package}
\alias{ohigui}
\alias{ohigui-package}
\title{Ocean Health Index - Graphical User Interface}
\description{
  The ohigui is an interface through the web browser for
  exploring and calculating the Ocean Health Index.
}
\author{
  Ben Best \email{bbest@nceas.ucsb.edu}
}

